README.TXT
Hello, I'm creating a DEFI platform on Litecoin for borrowing and lending purposes to enable individuals globally to participate in a sound financial system.
Offering the opportunity for people around the globe, will help promote an economical growth and lead cities and countries out of poverty. 
A financial system is a key focus on that growth and it must be done.